% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HTSvis.R
\name{HTSvis}
\alias{HTSvis}
\title{Launch the HTSvis app}
\usage{
HTSvis()
}
\description{
Executing this function will launch the HTSvis application in
the user's default web browser.
HTSvis is a shiny application for the visulaization and analysis
of data from arrayed high-throughput screens.
}
\examples{
\dontrun{
HTSvis()
}
}
\author{
Christian Scheeder \email{c.scheeder@dkfz.de}

Florian Heigwer \email{f.heigwer@dkfz.de}

Michael Boutros \email{m.boutros@dkfz.de}
}

